+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 # Changelog
 
+-   `0.4.0`
+
+    -   Modified `setup` to ask the user to select the default package manager to install
+        dependencies
+    -   Added `--yarn` flag to `setup` to use Yarn as the default package manager
+    -   Added `--npm` flag to `setup` to use NPM as the default package manager
+
+-   `0.3.3`
+
+    -   Finalization of project features
+
 -   `0.1.0`
 
     -   Added [core](src/core) files containing utility functions used by the features
@@ -13,7 +24,7 @@
         insert relationships between models. It also creates the migrations to commit changes to
         the database
     -   Added the [setup](src/features/setup.ts) feature. This processes the `model-definitions.js`
-        file at the project root and uses the [generate](src/features/generate.ts) and 
-        [link](src/features/link.ts) features to create all models and relations specified in the 
+        file at the project root and uses the [generate](src/features/generate.ts) and
+        [link](src/features/link.ts) features to create all models and relations specified in the
         file
     -   Added tests for all [core](src/core) and [feature](src/features) files